Moved proxy registry initialization to app setup stage
